Move memory processing logic in node.py into memory_parser
